# Presentation/Angular/src/app/api/models/invoice-per-month-list-result.ts
/* tslint:disable */
/* eslint-disable */
import { InvoicePerMonth } from './invoice-per-month';
export interface InvoicePerMonthListResult {
  data?: null | Array<InvoicePerMonth>;
  message?: null | string;
}