# Presentation/Angular/src/app/api/models/invoice-type-count-per-month-list-result.ts
/* tslint:disable */
/* eslint-disable */
import { InvoiceTypeCountPerMonth } from './invoice-type-count-per-month';
export interface InvoiceTypeCountPerMonthListResult {
  data?: null | Array<InvoiceTypeCountPerMonth>;
  message?: null | string;
}